✨ Update boost list messages and reactions 🚀
- Display active boosters and token sending guidance based on contract state.
- Provide clear instructions for boosting and joining the boost list.
- Update notifications for sending tokens based on contract state.
