+++ main.go
@@ -150,144 +150,10 @@ var (
 	}
 
 	commands = []*discordgo.ApplicationCommand{
-		{
-			Name:        slashContract,
-			Description: "Create a contract boost list.",
-			/*
-				Contexts: &[]discordgo.InteractionContextType{
-					discordgo.InteractionContextGuild,
-					discordgo.InteractionContextBotDM,
-					discordgo.InteractionContextPrivateChannel,
-				},
-				IntegrationTypes: &[]discordgo.ApplicationIntegrationType{
-					discordgo.ApplicationIntegrationGuildInstall,
-					discordgo.ApplicationIntegrationUserInstall,
-				},
-			*/
-			Options: []*discordgo.ApplicationCommandOption{
-				{
-					Type:         discordgo.ApplicationCommandOptionString,
-					Name:         "contract-id",
-					Description:  "Contract ID",
-					Required:     true,
-					Autocomplete: true,
-				},
-				{
-					Type:        discordgo.ApplicationCommandOptionString,
-					Name:        "coop-id",
-					Description: "Coop ID",
-					Required:    true,
-				},
-				{
-					Type:        discordgo.ApplicationCommandOptionInteger,
-					Name:        "coop-size",
-					Description: "Co-op Size. This will be pulled from EI Contract data if unset.",
-					Required:    false,
-				},
-				{
-					Name:        "boost-order",
-					Description: "Select how boost list is ordered. Default is Sign-up order.",
-					Required:    false,
-					Type:        discordgo.ApplicationCommandOptionInteger,
-					Choices: []*discordgo.ApplicationCommandOptionChoice{
-						{
-							Name:  "Sign-up Ordering",
-							Value: boost.ContractOrderSignup,
-						},
-						{
-							Name:  "Fair Ordering",
-							Value: boost.ContractOrderFair,
-						},
-						{
-							Name:  "Random Ordering",
-							Value: boost.ContractOrderRandom,
-						},
-					},
-				},
-				{
-					Type:        discordgo.ApplicationCommandOptionRole,
-					Name:        "ping-role",
-					Description: "Role to use to ping for this contract. Default is @here.",
-					Required:    false,
-				},
-				{
-					Type:        discordgo.ApplicationCommandOptionBoolean,
-					Name:        "make-thread",
-					Description: "Create a thread for this contract? (default: false)",
-					Required:    false,
-				},
-			},
-		},
-		{
-			Name:        slashSpeedrun,
-			Description: "Add speedrun features to a contract.",
-			Options: []*discordgo.ApplicationCommandOption{
-				{
-					Type:        discordgo.ApplicationCommandOptionUser,
-					Name:        "sink-crt",
-					Description: "The user to sink during CRT. Used for other sink parameters if those are missing.",
-					Required:    true,
-				},
-				{
-					Type:        discordgo.ApplicationCommandOptionString,
-					Name:        "sink-boosting",
-					Description: "Sink during boosting.",
-					Required:    false,
-				},
-				{
-					Type:        discordgo.ApplicationCommandOptionString,
-					Name:        "sink-post",
-					Description: "Post contract sink.",
-					Required:    false,
-				},
-				{
-					Name:        "sink-position",
-					Description: "Default is First Booster",
-					Required:    false,
-					Type:        discordgo.ApplicationCommandOptionInteger,
-					Choices: []*discordgo.ApplicationCommandOptionChoice{
-						{
-							Name:  "First",
-							Value: boost.SinkBoostFirst,
-						},
-						{
-							Name:  "Last",
-							Value: boost.SinkBoostLast,
-						},
-					},
-				},
-				{
-					Name:        "style",
-					Description: "Style of speedrun. Default is Banker",
-					Required:    false,
-					Type:        discordgo.ApplicationCommandOptionInteger,
-					Choices: []*discordgo.ApplicationCommandOptionChoice{
-						{
-							Name:  "Banker",
-							Value: boost.SpeedrunStyleBanker,
-						},
-						{
-							Name:  "Boost List",
-							Value: boost.SpeedrunStyleFastrun,
-						},
-					},
-				},
-				{
-					Type:        discordgo.ApplicationCommandOptionInteger,
-					Name:        "chicken-runs",
-					Description: "Number of chicken runs for this contract. Optional if contract-id was selected via auto fill.",
-					MinValue:    &integerZeroMinValue,
-					MaxValue:    20,
-					Required:    false,
-				},
-				{
-					Type:        discordgo.ApplicationCommandOptionBoolean,
-					Name:        "self-runs",
-					Description: "Self Runs during CRT",
-					Required:    false,
-				},
-			},
-		},
+
+		boost.GetSlashContractCommand(slashContract),
+		boost.GetSlashSpeedrunCommand(slashSpeedrun),
+
 		boost.GetSlashChangeSpeedRunSinkCommand(slashChangeSpeedRunSink),
 		{
 			Name:        slashJoin,
