+++ src/boost/boost_sink.go
@@ -84,12 +84,13 @@ func HandleSlashVolunteerSinkCommand(s *discordgo.Session, i *discordgo.Interact
 			str = "You must confirm you want to be the token sink"
 		} else if contract.Speedrun {
 			str = "You cannot use this command on a speedrun contract"
-		} else if contract.VolunteerSink != "" && !isAdmin {
+		} else if contract.Banker.VolunteerSink != "" && !isAdmin {
 			str = "Token sink is already set"
 		} else {
 			// Check if user is already in contract
 			if userInContract(contract, i.Interaction.Member.User.ID) {
-				contract.VolunteerSink = i.Interaction.Member.User.ID
+				contract.Banker.VolunteerSink = i.Interaction.Member.User.ID
+				changeContractState(contract, contract.State) // Update the changed sink
 				if contract.State == ContractStateCompleted || contract.State == ContractStateWaiting {
 					_ = RedrawBoostList(s, i.GuildID, i.ChannelID)
 				}
@@ -158,13 +159,14 @@ func HandleSlashVoluntellSinkCommand(s *discordgo.Session, i *discordgo.Interact
 			str = "This should be a guest farmer within this contract and not a user mention."
 		} else if contract.Speedrun {
 			str = "You cannot use this command on a speedrun contract"
-		} else if contract.VolunteerSink != "" && !isAdmin {
+		} else if contract.Banker.VolunteerSink != "" && !isAdmin {
 			str = "Token sink is already set"
 		} else {
 			// if VolunteerSink is already set, reply with error
 			// Check if user is already in contract
 			if userInContract(contract, VoluntellName) {
-				contract.VolunteerSink = VoluntellName
+				contract.Banker.VolunteerSink = VoluntellName
+				changeContractState(contract, contract.State) // Update the changed sink
 				if contract.State == ContractStateCompleted || contract.State == ContractStateWaiting {
 					_ = RedrawBoostList(s, i.GuildID, i.ChannelID)
 				}
