+++ src/boost/boost_slashcmd.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/bwmarrin/discordgo"
-	"github.com/mkmccarty/TokenTimeBoostBot/src/ei"
 	"github.com/mkmccarty/TokenTimeBoostBot/src/farmerstate"
 	"github.com/xhit/go-str2duration/v2"
 )
@@ -59,159 +58,6 @@ func UpdateThreadName(s *discordgo.Session, contract *Contract) {
 				})
 			}
 		}
-	}
-}
-
-// HandleContractCommand will handle the /contract command
-func HandleContractCommand(s *discordgo.Session, i *discordgo.InteractionCreate) {
-	// Protection against DM use
-	if i.GuildID == "" {
-		_ = s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-			Type: discordgo.InteractionResponseChannelMessageWithSource,
-			Data: &discordgo.InteractionResponseData{
-				Content:    "This command can only be run in a server.",
-				Flags:      discordgo.MessageFlagsEphemeral,
-				Components: []discordgo.MessageComponent{}},
-		})
-		return
-	}
-	var contractID = i.GuildID
-	var coopID = i.GuildID // Default to the Guild ID
-	var boostOrder = ContractOrderSignup
-	var coopSize = 0
-	var ChannelID = i.ChannelID
-	var pingRole = "@here"
-	makeThread := false
-
-	// User interacting with bot, is this first time ?
-	options := i.ApplicationCommandData().Options
-	optionMap := make(map[string]*discordgo.ApplicationCommandInteractionDataOption, len(options))
-	for _, opt := range options {
-		optionMap[opt.Name] = opt
-	}
-
-	if opt, ok := optionMap["coop-size"]; ok {
-		coopSize = int(opt.IntValue())
-	}
-	if opt, ok := optionMap["ping-role"]; ok {
-		role := opt.RoleValue(nil, "")
-		pingRole = role.Mention()
-	}
-	if opt, ok := optionMap["boost-order"]; ok {
-		boostOrder = int(opt.IntValue())
-	}
-	if opt, ok := optionMap["contract-id"]; ok {
-		contractID = opt.StringValue()
-		contractID = strings.Replace(contractID, " ", "", -1)
-	}
-	if opt, ok := optionMap["coop-id"]; ok {
-		coopID = opt.StringValue()
-		coopID = strings.Replace(coopID, " ", "", -1)
-	} else {
-		var c, err = s.Channel(ChannelID)
-		if err != nil {
-			coopID = c.Name
-		}
-	}
-	if opt, ok := optionMap["make-thread"]; ok {
-		makeThread = opt.BoolValue()
-	}
-
-	if coopSize == 0 {
-		found := false
-		for _, x := range ei.EggIncContracts {
-			if x.ID == contractID {
-				found = true
-			}
-		}
-		if !found {
-			_ = s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-				Type: discordgo.InteractionResponseChannelMessageWithSource,
-				Data: &discordgo.InteractionResponseData{
-					Content:    "Select a contract-id from the dropdown list.\nIf the contract-id list doesn't have your contract then supply a coop-size parameter.",
-					Flags:      discordgo.MessageFlagsEphemeral,
-					Components: []discordgo.MessageComponent{}},
-			})
-			return
-		}
-	}
-
-	// Create a new thread for this contract
-	if makeThread {
-		ch, err := s.Channel(ChannelID)
-		if err == nil {
-			if !ch.IsThread() {
-				// Default to 1 day timeout
-				var builder strings.Builder
-				builder.WriteString(coopID)
-				info := ei.EggIncContractsAll[contractID]
-				if info.ID != "" {
-					fmt.Fprintf(&builder, " (0/%d)", info.MaxCoopSize)
-				}
-
-				thread, err := s.ThreadStart(ChannelID, builder.String(), discordgo.ChannelTypeGuildPublicThread, 60*24)
-				if err == nil {
-					ChannelID = thread.ID
-					_ = s.ThreadJoin(getInteractionUserID(i))
-				} else {
-					log.Print(err)
-				}
-			}
-		}
-	}
-
-	mutex.Lock()
-	contract, err := CreateContract(s, contractID, coopID, coopSize, boostOrder, i.GuildID, ChannelID, getInteractionUserID(i), pingRole)
-	mutex.Unlock()
-	if err != nil {
-
-		_ = s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-			Type: discordgo.InteractionResponseChannelMessageWithSource,
-			Data: &discordgo.InteractionResponseData{
-				Content:    err.Error(),
-				Flags:      discordgo.MessageFlagsEphemeral,
-				Components: []discordgo.MessageComponent{}},
-		})
-		return
-	}
-
-	var builder strings.Builder
-	builder.WriteString("Contract created. Use the Contract button if you have to recycle it.\n")
-	builder.WriteString("This fastrun contract can be converted to a `/speedrun` anytime during the sign-up list.\n")
-	builder.WriteString("If this contract isn't an immediate start use `/change-planned-start` to add the time to the sign-up message.\n")
-	builder.WriteString("React with 🌊 to automaticaly update the thread name.")
-
-	err = s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-		Type: discordgo.InteractionResponseChannelMessageWithSource,
-		Data: &discordgo.InteractionResponseData{
-			Content: builder.String(),
-			Flags:   discordgo.MessageFlagsEphemeral,
-			// Buttons and other components are specified in Components field.
-		},
-	})
-	if err != nil {
-		log.Print(err)
-	}
-
-	var createMsg = DrawBoostList(s, contract)
-	var data discordgo.MessageSend
-	data.Content = createMsg
-	data.Flags = discordgo.MessageFlagsSuppressEmbeds
-	msg, err := s.ChannelMessageSendComplex(ChannelID, &data)
-	if err == nil {
-		SetListMessageID(contract, ChannelID, msg.ID)
-		var data discordgo.MessageSend
-		data.Content, data.Components = GetSignupComponents(false, contract.Speedrun)
-		reactionMsg, err := s.ChannelMessageSendComplex(ChannelID, &data)
-
-		if err != nil {
-			log.Print(err)
-		} else {
-			SetReactionID(contract, msg.ChannelID, reactionMsg.ID)
-			_ = s.ChannelMessagePin(msg.ChannelID, reactionMsg.ID)
-		}
-	} else {
-		log.Print(err)
 	}
 }
